+++ packages/firebase_ml_vision/ios/Classes/FirebaseMlVisionPlugin.h
@@ -26,5 +26,5 @@
 @interface CloudLabelDetector : NSObject<Detector>
 @end
 
-@interface TextDetector : NSObject<Detector>
+@interface TextRecognizer : NSObject<Detector>
 @end