+++ packages/firebase_ml_vision/ios/Classes/FirebaseMlVisionPlugin.m
@@ -46,8 +46,8 @@
     [LabelDetector handleDetection:image options:options result:result];
   } else if ([@"CloudLabelDetector#detectInImage" isEqualToString:call.method]) {
     [CloudLabelDetector handleDetection:image options:options result:result];
-  } else if ([@"TextDetector#detectInImage" isEqualToString:call.method]) {
-    [TextDetector handleDetection:image options:options result:result];
+  } else if ([@"TextRecognizer#detectInImage" isEqualToString:call.method]) {
+    [TextRecognizer handleDetection:image options:options result:result];
   } else {
     result(FlutterMethodNotImplemented);
   }