+++ packages/firebase_storage/lib/src/upload_task.dart
@@ -6,62 +6,160 @@ part of firebase_storage;
 
 abstract class StorageUploadTask {
   final FirebaseStorage _firebaseStorage;
-  final String _path;
+  final StorageReference _ref;
   final StorageMetadata _metadata;
 
-  StorageUploadTask._(this._firebaseStorage, this._path, this._metadata);
-  Future<void> _start();
+  StorageUploadTask._(this._firebaseStorage, this._ref, this._metadata);
+  Future<dynamic> _platformStart();
 
-  Completer<UploadTaskSnapshot> _completer =
-      new Completer<UploadTaskSnapshot>();
-  Future<UploadTaskSnapshot> get future => _completer.future;
+  int _handle;
+
+  bool isCanceled = false;
+  bool isComplete = false;
+  bool isInProgress = true;
+  bool isPaused = false;
+  bool isSuccessful = false;
+
+  StorageTaskSnapshot lastSnapshot;
+
+  /// Returns a last snapshot when completed
+  Completer<StorageTaskSnapshot> _completer =
+      new Completer<StorageTaskSnapshot>();
+  Future<StorageTaskSnapshot> get onComplete => _completer.future;
+
+  StreamController<StorageTaskEvent> _controller =
+      new StreamController<StorageTaskEvent>.broadcast();
+  Stream<StorageTaskEvent> get events => _controller.stream;
+
+  Future<StorageTaskSnapshot> _start() async {
+    _handle = await _platformStart();
+    final StorageTaskEvent event = await _firebaseStorage._methodStream
+        .where((MethodCall m) {
+      return m.method == 'StorageTaskEvent' && m.arguments['handle'] == _handle;
+    }).map<StorageTaskEvent>((MethodCall m) {
+      final Map<dynamic, dynamic> args = m.arguments;
+      final StorageTaskEvent e =
+          new StorageTaskEvent._(args['type'], _ref, args['snapshot']);
+      _changeState(e);
+      lastSnapshot = e.snapshot;
+      _controller.add(e);
+      if (e.type == StorageTaskEventType.success ||
+          e.type == StorageTaskEventType.failure) {
+        _completer.complete(e.snapshot);
+      }
+      return e;
+    }).firstWhere((StorageTaskEvent e) =>
+            e.type == StorageTaskEventType.success ||
+            e.type == StorageTaskEventType.failure);
+    return event.snapshot;
+  }
+
+  void _changeState(StorageTaskEvent event) {
+    _resetState();
+    print('EVENT ${event.type}');
+    switch (event.type) {
+      case StorageTaskEventType.progress:
+        isInProgress = true;
+        break;
+      case StorageTaskEventType.resume:
+        isInProgress = true;
+        break;
+      case StorageTaskEventType.pause:
+        isPaused = true;
+        break;
+      case StorageTaskEventType.success:
+        isSuccessful = true;
+        isComplete = true;
+        break;
+      case StorageTaskEventType.failure:
+        isComplete = true;
+        if (event.snapshot.error == StorageError.canceled) {
+          isCanceled = true;
+        }
+        break;
+    }
+  }
+
+  void _resetState() {
+    isCanceled = false;
+    isComplete = false;
+    isInProgress = false;
+    isPaused = false;
+    isSuccessful = false;
+  }
+
+  /// Pause the upload
+  void pause() => FirebaseStorage.channel.invokeMethod(
+        'UploadTask#pause',
+        <String, dynamic>{
+          'app': _firebaseStorage.app?.name,
+          'bucket': _firebaseStorage.storageBucket,
+          'handle': _handle,
+        },
+      );
+
+  /// Resume the upload
+  void resume() => FirebaseStorage.channel.invokeMethod(
+        'UploadTask#resume',
+        <String, dynamic>{
+          'app': _firebaseStorage.app?.name,
+          'bucket': _firebaseStorage.storageBucket,
+          'handle': _handle,
+        },
+      );
+
+  /// Cancel the upload
+  void cancel() => FirebaseStorage.channel.invokeMethod(
+        'UploadTask#cancel',
+        <String, dynamic>{
+          'app': _firebaseStorage.app?.name,
+          'bucket': _firebaseStorage.storageBucket,
+          'handle': _handle,
+        },
+      );
 }
 
 class _StorageFileUploadTask extends StorageUploadTask {
   final File _file;
   _StorageFileUploadTask._(this._file, FirebaseStorage firebaseStorage,
-      String path, StorageMetadata metadata)
-      : super._(firebaseStorage, path, metadata);
+      StorageReference ref, StorageMetadata metadata)
+      : super._(firebaseStorage, ref, metadata);
 
   @override
-  Future<void> _start() async {
-    final String downloadUrl = await FirebaseStorage.channel.invokeMethod(
+  Future<dynamic> _platformStart() {
+    return FirebaseStorage.channel.invokeMethod(
       'StorageReference#putFile',
       <String, dynamic>{
         'app': _firebaseStorage.app?.name,
         'bucket': _firebaseStorage.storageBucket,
         'filename': _file.absolute.path,
-        'path': _path,
+        'path': _ref.path,
         'metadata':
             _metadata == null ? null : _buildMetadataUploadMap(_metadata),
       },
     );
-    _completer
-        .complete(new UploadTaskSnapshot(downloadUrl: Uri.parse(downloadUrl)));
   }
 }
 
 class _StorageDataUploadTask extends StorageUploadTask {
   final Uint8List _bytes;
   _StorageDataUploadTask._(this._bytes, FirebaseStorage firebaseStorage,
-      String path, StorageMetadata metadata)
-      : super._(firebaseStorage, path, metadata);
+      StorageReference ref, StorageMetadata metadata)
+      : super._(firebaseStorage, ref, metadata);
 
   @override
-  Future<void> _start() async {
-    final String downloadUrl = await FirebaseStorage.channel.invokeMethod(
+  Future<dynamic> _platformStart() {
+    return FirebaseStorage.channel.invokeMethod(
       'StorageReference#putData',
       <String, dynamic>{
         'app': _firebaseStorage.app?.name,
         'bucket': _firebaseStorage.storageBucket,
         'data': _bytes,
-        'path': _path,
+        'path': _ref.path,
         'metadata':
             _metadata == null ? null : _buildMetadataUploadMap(_metadata),
       },
     );
-    _completer
-        .complete(new UploadTaskSnapshot(downloadUrl: Uri.parse(downloadUrl)));
   }
 }
 
@@ -74,9 +172,4 @@ Map<String, dynamic> _buildMetadataUploadMap(StorageMetadata metadata) {
     'contentEncoding': metadata.contentEncoding,
     'customMetadata': metadata.customMetadata,
   };
-}
-
-class UploadTaskSnapshot {
-  UploadTaskSnapshot({this.downloadUrl});
-  final Uri downloadUrl;
 }