+++ packages/cloud_functions/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.4
+
+* Fixed podspec to use static_framework
+
+## 0.0.3
+
+* Added missing dependency on meta package.
+
 ## 0.0.2
 
 * Bump Android and Firebase dependency versions.