+++ packages/firebase_database/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.4
+
+* Bumped test and mockito versions to pick up Dart 2 support.
+
 ## 1.0.3
 
 * Bump Android and Firebase dependency versions.