+++ packages/firebase_messaging/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.0.0
+
+* Updated Android to send Remote Message's title and body to Dart.
+
+## 1.0.5
+
+* Bumped test and mockito versions to pick up Dart 2 support.
+
 ## 1.0.4
 
 * Bump Android and Firebase dependency versions.